requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
